+++ tests/hazmat/primitives/test_xofhash.py
@@ -78,10 +78,8 @@ class TestXOFHash:
             actual = hashes.XOFHash(algorithm.xof())
             actual.update(data)
 
-            assert (
-                actual.squeeze(16) + actual.squeeze(16)
-                == expected.squeeze(32)
-            )
+            actual_output = actual.squeeze(16) + actual.squeeze(16)
+            assert actual_output == expected.squeeze(32)
 
     def test_exhaust_bytes(self):
         h = hashes.XOFHash(hashes.SHAKE128(digest_size=256))